fix: line and column ordinals are u64, and never clamped
`Position` said "this is where it is" and `saturating_u32` said "this might
have been clamped and you cannot tell". A consumer could not distinguish a
real line 4294967295 from a saturated one, which is a narrowing that lies
rather than one that fails.

The scenario behind it is not worth fixing on its own — `u32::MAX` line breaks
is four gigabytes of source, and a renderer will not see one. The API is. This
is the one type every renderer and every FFI consumer touches, its width is
unchangeable once published, and painty is unpublished, so it is free now and
a breaking change later.

`u64` rather than `usize`, because the resolved model is meant to cross a C
ABI and `usize` is platform-dependent. `u64` also makes the question disappear
instead of moving it: an ordinal is at most one more than the text's length in
bytes, Rust caps an object at `isize::MAX`, so it cannot exceed 2^63 and
cannot overflow. The increments are therefore plain `+` — a debug build panics
if that reasoning is ever wrong, which is the direction to fail in.

Two alternatives were considered and rejected. An explicit overflow state
makes every consumer handle a case that cannot occur. Keeping `u32` and
refusing an oversized source at construction is honest, but it puts a fallible
constructor on the common path to guard an input nobody has.

APPLIED BY CATEGORY, not field by field, because a model where two of five
widened is worse than either uniform choice. Every line or column ordinal —
`Position::line`, `Position::column`, `Line::number`, `Line::char_count`,
`Line::column_at`, `Source::line_count`, `Source::line`, `Region::line_count`,
`RegionLine::columns` — is a `u64`. Every byte offset — `Span`,
`Position::offset`, `Line::span` — stays `usize`, because it indexes the
caller's `&str` and that is what Rust slices with, it is exact, and the
widening to `u64` an FFI export needs is lossless from every `usize` Rust has.
`Location::source` also stays `u32`: painty never computes it, so it cannot
lie, and it is a key into the caller's own list that has to round-trip through
a contract already spelling it that way.

`Source::line_count` now reads the last line's ordinal instead of counting
into a `usize` and converting, and `RegionLines::size_hint` — the one place
the two units genuinely meet — uses `try_from` and answers "unknown" rather
than a wrong lower bound, which `Vec`'s `Extend` reserves from.

Costs eight bytes per ordinal on stack-only `Copy` structs handled a handful
of times per diagnostic: `Position` 24, `Line` 40, `Region` 96, `RegionLine`
56. Nothing here allocates, so there is nothing else to pay.

+++ src/source/tests.rs
@@ -2,7 +2,7 @@ use super::{LineBreak, Source};
 use crate::Span;
 
 /// The line's number, its text, and what ended it.
-fn shape(line: crate::Line<'_>) -> (u32, &str, Option<LineBreak>) {
+fn shape(line: crate::Line<'_>) -> (u64, &str, Option<LineBreak>) {
   (line.number(), line.text(), line.line_break())
 }
